Add files via upload
We have identified the key parameters and determined the target attribute (CTR) that we would like to predict based on the ad parameters. At the same time, we want to build models periodically based on new data and use it over incoming data, so we will need two Spark tasks to perform.